Added Desert Map
Created the desert map for level 2, it is much more open than the last one because, you know, deserts are empty. Also I struggled finding a good tilemap for free for it, so I just ripped off the original Legend of Zelda. There isn't a canopy layer this time since there wasn't anything to put above the player